requirements.txt
behave
#pymongo
pyyaml
elasticsearch
Appium-Python-Client
redis
sure
uuid
celery
requests
selenium
apscheduler
suds
suds_requests
